requirements.txt
aiohttp==3.9.3
discord.py[voice]==2.3.2
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
langchain-core==0.1.50
langchain-openai==0.1.6
langchain-community==0.0.34
langchain-chroma==0.1.0
openai==1.23.2
pydantic==2.7.1
python-dotenv==1.0.1
